+++ previous-versions/assess/RELEASES.md
@@ -657,3 +657,13 @@ Based on:
 - [typescript v0.37.0] previous-versions/assess
 ### Releases
 - [NPM v0.37.0] https://www.npmjs.com/package/@codat/assess/v/0.37.0 - previous-versions/assess
+
+## 2023-10-02 00:21:10
+### Changes
+Based on:
+- OpenAPI Doc 1.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Assess.yaml
+- Speakeasy CLI 1.92.3 (2.143.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.37.1] previous-versions/assess
+### Releases
+- [NPM v0.37.1] https://www.npmjs.com/package/@codat/assess/v/0.37.1 - previous-versions/assess